requirements/base.txt
python-dotenv
aiofiles==24.1.0
aiogram==3.13.1
aiohttp
aioredis
asyncpg
